Move InstancesFullEditor to SceneEditor
