+++ newIDE/app/src/SceneEditor/InstancesSelection.js
