+++ newIDE/app/src/SceneEditor/ScenePropertiesDialog.js
@@ -2,11 +2,11 @@ import React, { Component } from 'react';
 import FlatButton from 'material-ui/FlatButton';
 import TextField from 'material-ui/TextField';
 import RaisedButton from 'material-ui/RaisedButton';
-import Dialog from '../../UI/Dialog';
-import ColorField from '../../UI/ColorField';
-import EmptyMessage from '../../UI/EmptyMessage';
-import PropertiesEditor from '../../PropertiesEditor';
-import propertiesMapToSchema from '../../PropertiesEditor/PropertiesMapToSchema';
+import Dialog from '../UI/Dialog';
+import ColorField from '../UI/ColorField';
+import EmptyMessage from '../UI/EmptyMessage';
+import PropertiesEditor from '../PropertiesEditor';
+import propertiesMapToSchema from '../PropertiesEditor/PropertiesMapToSchema';
 import some from 'lodash/some';
 
 export default class ScenePropertiesDialog extends Component {
